requirements.txt
flask==2.2.3
werkzeug==2.2.3
sqlalchemy==2.0.4
flask-migrate==4.0.4
psycopg2-binary==2.9.5
python-dotenv==1.0.0
beautifulsoup4==4.11.2
requests==2.28.2
alembic==1.10.2
flasgger==0.9.7.1
apscheduler==3.10.1